requirements.txt
pypdf
langchain
langchain-community>=0.3.7
langchain-chroma
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
sentence-transformers
timedelta
langchain_huggingface